予約作成途中

+++ LibraryManagement/src/main/java/app/repository/MemberRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import app.entity.Member;
 
 @Repository
-public interface MemberRepository extends JpaRepository<Member, Long> {
-    boolean existsByEmail(String email);  // メールの重複チェック用（任意）
+public interface MemberRepository extends JpaRepository<Member, Integer> {
+	boolean existsByEmail(String email);  // メールの重複チェック用（任意）
 }
